stdlib: Checking argument against io:format()
As of commit 75e0c601890d9561b103936479de0483cbcafd9d OTP is free of
"bad" calls (a call is bad if the format is not an io:format()). This
branch is to ensure it stays that way.

It is not yet decided if (an improved version) of the check will
be included in Erlang/OTP 22.