Add module hello as a minimal test of JIT for ARM
This commit modifies the runtime system so that the hello
will be loaded as one of the pre-loaded modules, and so
that hello:start/2 will be called.

How to use this commit:

First build OTP with this commit reverted, and with the JIT
disabled:

    ./otp_build setup --disable-jit

Then unrevert this commit and execute the following commands:

    rm $ERL_TOP/erts/emulator/*/opt/jit/*GENERATED
    ./otp_build autoconf && ./configure
    make emulator
    bin/erl

If everything is working, the output should be similar to:

    hello, world
    arguments: [<<"-root">>,<<"/Users/bjorng/git/otp">>,<<"-progname">>,<<"erl">>,<<"--">>,<<"-home">>,<<"/Users/bjorng">>,<<"--">>,<<"-kernel">>,<<"shell_history">>,<<"enabled">>,<<"--">>]
    Testing stuff (should not crash)...
    Everything is fine!

Then a version of the EStone benchmark will be run.

The exit status should be 42:

    $ echo $?

Co-authored-by: John Högberg <john@erlang.org>
Co-authored-by: Sverker Eriksson <sverker@erlang.org>