erts: Accept faulty Size field in NEW_FUN_EXT
This is a pragmatic semi-solution to remove the symptom of a bug
that could cause invalid Size field in NEW_FUN_EXT encoded toward
a pending connection.
The Size field is not used for anything else than this sanitity check
so I see it safe to accept NEW_FUN_EXT from buggy senders.

erl_interface still relies on Size so I guess we have to fix the
encoding bug also, even though sending funs to erl_interface is not
that common.